+++ src/components/CreatePost/CaseStyle.js
